requirements.txt
Flask
flask_socketio
minimalmodbus
pyserial